task 02 added

+++ Task-02/src/task_interface/IResourceContext.java
@@ -1,6 +1,8 @@
 package task_interface;
 
+import java.math.BigInteger;
+
 public interface IResourceContext {
-    double FreeMemory();
-    double TotalMemory();
+    BigInteger FreeMemory();
+    BigInteger TotalMemory();
 }
